finish case-sensitive renaming
