+++ docs/components/shell/index.ts
